_requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
asgiref==3.6.0
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
av==10.0.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.10
boto3==1.29.6
botocore==1.32.6
breadability==0.1.20
cachetools==5.3.2
catalogue==2.0.9
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
chex==0.1.85
click==8.1.3
coloredlogs==15.0.1
comm==0.1.4
confection==0.1.1
cryptography==41.0.4
ctranslate2==3.20.0
cymem==2.0.7
debugpy==1.8.0
decorator==5.1.1
deepgram-sdk==2.11.0
distlib==0.3.6
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
etils==1.5.2
executing==2.0.0
fastapi==0.104.0
faster-whisper==0.9.0
filelock==3.12.0
Flask==2.3.1
Flask-Cors==3.0.10
flatbuffers==23.5.26
flax==0.7.5
frozenlist==1.4.0
fsspec==2023.9.2
gensim==4.3.2
gensim-sum-ext==0.1.2
gitdb==4.0.11
GitPython==3.1.40
google==3.0.0
googlesearch-python==1.2.3
h11==0.14.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.16.1
itsdangerous==2.1.2
jax==0.4.20
jaxlib==0.4.20
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.5.0
jupyter_core==5.4.0
langcodes==3.3.0
llvmlite==0.41.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.1
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multipledispatch==0.6.0
murmurhash==1.0.9
mutagen @ file:///private/tmp/python-mutagen-20231008-5554-1csby1/mutagen-1.47.0
ndg-httpsclient==0.5.1
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
numba==0.58.0
numpy==1.24.2
onnxruntime==1.16.1
openai-whisper==20230918
opencv-python==4.8.1.78
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.3
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.0
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.5.1
preshed==3.0.8
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.0
pycorenlp==0.3.0
pycountry==22.3.5
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.16.1
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.0
regex==2023.8.8
requests==2.31.0
rich==13.7.0
rpds-py==0.13.1
s3transfer==0.7.0
sacremoses==0.1.1
safetensors==0.4.0
scipy==1.11.4
selenium==4.14.0
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.10.0
srsly==2.4.7
stack-data==0.6.3
starlette==0.27.0
streamlit==1.28.2
summa==1.2.0
summarizer==0.0.7
sumy==0.11.0
sympy==1.12
tenacity==8.2.3
tensorstore==0.1.50
textsummarizer==1.0.0
thinc==8.1.12
tiktoken==0.3.3
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
transformers==4.11.3
trio==0.22.2
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.4
uvicorn==0.23.2
validators==0.22.0
virtualenv==20.23.0
wasabi==1.1.2
wcwidth==0.2.8
websockets==12.0
Werkzeug==2.3.1
wsproto==1.2.0
yarl==1.9.2
youtube-dl==2021.12.17
youtube-transcript-api==0.6.1
zipp==3.17.0
